Enable user to input customized text
Add customized input function in Response page. This added item does not
affect the vision page. Will remain for all this run (hold in memory,
not in disk)
Fix a bug, when come to response page from the root level, if click home
button, will go to home again instead of hearing